requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
camelot-py==1.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==43.0.3
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.40.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
gevent==25.5.1
ghostscript==0.8.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
greenlet==3.2.3
grequests==0.7.0
gspread==6.1.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector==2.2.9
numpy==2.3.0
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdfminer.six==20250506
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.2.3
pypdf==5.6.0
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyuwsgi==2.0.29
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.24.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uWSGI==2.0.30
waitress==3.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2